// ui/src/lib/generated/ApiErrorBody.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ErrorCode } from "./ErrorCode";

/**
 * エラー本体。
 */
export type ApiErrorBody = { 
/**
 * 構造化エラーコード
 */
code: ErrorCode, 
/**
 * 人間向けメッセージ
 */
message: string, };